refactor: rename introduction screens
